repositorios creados

+++ src/main/java/com/valencia/proyecto1evaluacion/repositorio/ProductoRepository.java
@@ -4,6 +4,8 @@ import com.valencia.proyecto1evaluacion.modelos.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 }
